% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tacos_cost.R
\name{tacos_cost}
\alias{tacos_cost}
\title{Calculate total taco cost}
\usage{
tacos_cost(price, count)
}
\arguments{
\item{price}{Price per taco in dollars}

\item{count}{Number of burritos in order}
}
\description{
Calculate total taco cost
}
